adopt allSettled in AsyncEmitter and move back into event.ts, https://github.com/microsoft/vscode/issues/121144
